+++ core/src/main/java/net/sourceforge/html5val/performers/MaxPerformer.java
@@ -1,7 +1,6 @@
 package net.sourceforge.html5val.performers;
 
 import javax.validation.constraints.Max;
-import net.sourceforge.html5val.ValidationPerformer;
 import org.thymeleaf.dom.Element;
 
 public class MaxPerformer implements ValidationPerformer<Max> {
