+++ core/src/main/java/net/sourceforge/html5val/performers/NotNullPerformer.java
@@ -1,6 +1,5 @@
 package net.sourceforge.html5val.performers;
 
-import net.sourceforge.html5val.ValidationPerformer;
 import org.thymeleaf.dom.Element;
 
 import javax.validation.constraints.NotNull;
