--- core/src/main/java/net/sourceforge/html5val/performers/RegexpComposer.java
@@ -1,7 +1,0 @@
-package net.sourceforge.html5val.performers;
-
-public interface RegexpComposer {
-
-    String regexp();
-
-}
