+++ core/src/main/java/net/sourceforge/html5val/performers/SizePerformer.java
@@ -1,9 +1,9 @@
 package net.sourceforge.html5val.performers;
 
-import net.sourceforge.html5val.ValidationPerformer;
 import org.thymeleaf.dom.Element;
 
 import javax.validation.constraints.Size;
+import net.sourceforge.html5val.performers.regexp_composer.LengthRegexpComposer;
 
 public class SizePerformer implements ValidationPerformer<Size> {
