+++ core/src/main/java/net/sourceforge/html5val/performers/URLPerformer.java
@@ -1,8 +1,8 @@
 package net.sourceforge.html5val.performers;
 
-import net.sourceforge.html5val.ValidationPerformer;
 import org.hibernate.validator.constraints.URL;
 import org.thymeleaf.dom.Element;
+import net.sourceforge.html5val.performers.regexp_composer.URLRegexpComposer;
 
 public class URLPerformer implements ValidationPerformer<URL> {
