+++ core/src/main/java/net/sourceforge/html5val/performers/ValidationPerformer.java
@@ -1,4 +1,4 @@
-package net.sourceforge.html5val;
+package net.sourceforge.html5val.performers;
 
 import java.lang.annotation.Annotation;
 import org.thymeleaf.dom.Element;
