+++ core/src/main/java/net/sourceforge/html5val/performers/regexp_composer/RegexpComposer.java
@@ -1,0 +1,7 @@
+package net.sourceforge.html5val.performers.regexp_composer;
+
+public interface RegexpComposer {
+
+    String regexp();
+
+}
