+++ core/src/main/java/net/sourceforge/html5val/reflect/FieldFinder.java
@@ -1,4 +1,4 @@
-package net.sourceforge.html5val;
+package net.sourceforge.html5val.reflect;
 
 import java.lang.reflect.Field;
