+++ core/src/main/java/net/sourceforge/html5val/thymeleaf3/FormSelectFinder.java
@@ -1,9 +1,11 @@
-package net.sourceforge.html5val;
+package net.sourceforge.html5val.thymeleaf3;
 
 import org.thymeleaf.dom.Element;
 
 import java.util.ArrayList;
 import java.util.List;
+
+import net.sourceforge.html5val.util.EmptyChecker;
 
 public class FormSelectFinder {
