+++ core/src/main/java/net/sourceforge/html5val/util/EmptyChecker.java
@@ -1,4 +1,4 @@
-package net.sourceforge.html5val;
+package net.sourceforge.html5val.util;
 
 import java.util.Collection;
