+++ core/src/test/java/unit/net/sourceforge/html5val/performers/LengthPerformerTest.java
@@ -1,6 +1,6 @@
 package unit.net.sourceforge.html5val.performers;
 
-import net.sourceforge.html5val.ValidationPerformer;
+import net.sourceforge.html5val.performers.ValidationPerformer;
 import net.sourceforge.html5val.performers.LengthPerformer;
 import org.hibernate.validator.constraints.Length;
 import org.junit.Before;
